+++ plugins/rmanipulation/basemanipulation.cpp
@@ -357,130 +357,17 @@ protected:
     bool MoveManipulator(ostream& sout, istream& sinput)
     {
         RAVELOG_DEBUG("Starting MoveManipulator...\n");
-        RobotBase::ManipulatorPtr pmanip = robot->GetActiveManipulator();
-
-        string strtrajfilename;
-        bool bExecute = true;
-        boost::shared_ptr<ostream> pOutputTrajStream;
-        std::vector<dReal> goals;
-        PlannerBase::PlannerParametersPtr params(new PlannerBase::PlannerParameters());
-        params->_nMaxIterations = 4000;     // max iterations before failure
-
-        string cmd;
-        int nMaxTries = 3;     // max tries for the planner
-        while(!sinput.eof()) {
-            sinput >> cmd;
-            if( !sinput ) {
-                break;
-            }
-            std::transform(cmd.begin(), cmd.end(), cmd.begin(), ::tolower);
-
-            if(( cmd == "armvals") ||( cmd == "goal") ) {
-                goals.resize(pmanip->GetArmIndices().size());
-                FOREACH(it, goals) {
-                    sinput >> *it;
-                }
-            }
-            else if( cmd == "outputtraj" ) {
-                pOutputTrajStream = boost::shared_ptr<ostream>(&sout,null_deleter());
-            }
-            else if( cmd == "maxiter" ) {
-                sinput >> params->_nMaxIterations;
-            }
-            else if( cmd == "execute" ) {
-                sinput >> bExecute;
-            }
-            else if( cmd == "writetraj" ) {
-                sinput >> strtrajfilename;
-            }
-            else if( cmd == "maxtries" ) {
-                sinput >> nMaxTries;
-            }
-            else {
-                RAVELOG_WARN(str(boost::format("unrecognized command: %s\n")%cmd));
-                break;
-            }
-
-            if( !sinput ) {
-                RAVELOG_ERROR(str(boost::format("failed processing command %s\n")%cmd));
-                return false;
-            }
-        }
-
-        if( goals.size() != pmanip->GetArmIndices().size() ) {
-            return false;
-        }
-
-        RobotBase::RobotStateSaver saver(robot);
-
-        robot->SetActiveDOFs(pmanip->GetArmIndices());
-        params->SetRobotActiveJoints(robot);
-
-        CollisionOptionsStateSaver optionstate(GetEnv()->GetCollisionChecker(),GetEnv()->GetCollisionChecker()->GetCollisionOptions()|CO_ActiveDOFs,false);
-
-        TrajectoryBasePtr ptraj = RaveCreateTrajectory(GetEnv(),robot->GetActiveDOF());
-
-        std::vector<dReal> values;
-        robot->GetActiveDOFValues(values);
-
-        // make sure the initial and goal configs are not in collision
-        robot->SetActiveDOFValues(goals, true);
-        if( planningutils::JitterActiveDOF(robot) == 0 ) {
-            RAVELOG_WARN("jitter failed\n");
-            return false;
-        }
-
-        robot->GetActiveDOFValues(params->vgoalconfig);
-        robot->SetActiveDOFValues(values);
-
-        // jitter again for initial collision
-        if( planningutils::JitterActiveDOF(robot) == 0 ) {
-            RAVELOG_WARN("jitter failed\n");
-            return false;
-        }
-        robot->GetActiveDOFValues(params->vinitialconfig);
-
-        PlannerBasePtr rrtplanner = RaveCreatePlanner(GetEnv(),_strRRTPlannerName);
-        if( !rrtplanner ) {
-            RAVELOG_WARN("failed to create planner\n");
-            return false;
-        }
-
-        bool bSuccess = false;
-        RAVELOG_INFO("starting planning\n");
-
-        for(int iter = 0; iter < nMaxTries; ++iter) {
-            if( !rrtplanner->InitPlan(robot, params) ) {
-                RAVELOG_ERROR("InitPlan failed\n");
-                break;
-            }
-
-            if( rrtplanner->PlanPath(ptraj) ) {
-                bSuccess = true;
-                RAVELOG_INFO("finished planning\n");
-                break;
-            }
-            else {
-                RAVELOG_WARN("PlanPath failed\n");
-            }
-        }
-
-        if( !bSuccess ) {
-            return false;
-        }
-        if( RaveGetDebugLevel() & Level_VerifyPlans ) {
-            planningutils::VerifyTrajectory(params, ptraj);
-        }
-        CM::SetActiveTrajectory(robot, ptraj, bExecute, strtrajfilename, pOutputTrajStream,_fMaxVelMult);
-        sout << "1";
-        return true;
+        RobotBase::RobotStateSaver saver(robot,KinBody::Save_ActiveDOF);
+        robot->SetActiveDOFs(robot->GetActiveManipulator()->GetArmIndices());
+        BOOST_ASSERT(robot->GetActiveDOF()>0);
+        return MoveActiveJoints(sout,sinput);
     }
 
     bool MoveActiveJoints(ostream& sout, istream& sinput)
     {
         string strtrajfilename;
         bool bExecute = true;
-        int nMaxTries = 1;     // max tries for the planner
+        int nMaxTries = 2;     // max tries for the planner
         boost::shared_ptr<ostream> pOutputTrajStream;
 
         PlannerBase::PlannerParametersPtr params(new PlannerBase::PlannerParameters());
@@ -495,9 +382,21 @@ protected:
             std::transform(cmd.begin(), cmd.end(), cmd.begin(), ::tolower);
 
             if( cmd == "goal" ) {
-                params->vgoalconfig.resize(robot->GetActiveDOF());
-                FOREACH(it, params->vgoalconfig) {
-                    sinput >> *it;
+                params->vgoalconfig.reserve(params->vgoalconfig.size()+robot->GetActiveDOF());
+                for(int i = 0; i < robot->GetActiveDOF(); ++i) {
+                    dReal f=0;
+                    sinput >> f;
+                    params->vgoalconfig.push_back(f);
+                }
+            }
+            else if( cmd == "goals" ) {
+                size_t numgoals = 0;
+                sinput >> numgoals;
+                params->vgoalconfig.reserve(params->vgoalconfig.size()+numgoals*robot->GetActiveDOF());
+                for(size_t i = 0; i < numgoals*robot->GetActiveDOF(); ++i) {
+                    dReal f=0;
+                    sinput >> f;
+                    params->vgoalconfig.push_back(f);
                 }
             }
             else if( cmd == "outputtraj" ) {
@@ -529,27 +428,45 @@ protected:
             }
         }
 
-        if( (int)params->vgoalconfig.size() != robot->GetActiveDOF() ) {
+        if( params->vgoalconfig.size() == 0 ) {
             return false;
         }
         RobotBase::RobotStateSaver saver(robot);
+        params->SetRobotActiveJoints(robot);
+
         CollisionOptionsStateSaver optionstate(GetEnv()->GetCollisionChecker(),GetEnv()->GetCollisionChecker()->GetCollisionOptions()|CO_ActiveDOFs,false);
 
-        if( planningutils::JitterActiveDOF(robot) == 0 ) {
-            RAVELOG_WARN("failed\n");
+        // make sure the initial and goal configs are not in collision
+        std::vector<dReal> originalvalues;
+        robot->GetActiveDOFValues(originalvalues);
+        vector<dReal> vonegoal(robot->GetActiveDOF());
+
+        size_t writeindex=0;
+        for(size_t i = 0; i < params->vgoalconfig.size(); i += robot->GetActiveDOF()) {
+            std::copy(params->vgoalconfig.begin()+i,params->vgoalconfig.begin()+i+robot->GetActiveDOF(),vonegoal.begin());
+            robot->SetActiveDOFValues(vonegoal, true);
+            if( planningutils::JitterActiveDOF(robot) == 0 ) {
+                RAVELOG_WARN(str(boost::format("jitter failed %d\n")%i));
+            }
+            else {
+                robot->GetActiveDOFValues(vonegoal);
+                std::copy(vonegoal.begin(),vonegoal.end(),params->vgoalconfig.begin()+writeindex);
+                writeindex += robot->GetActiveDOF();
+            }
+        }
+        if( writeindex == 0 ) {
             return false;
         }
+        params->vgoalconfig.resize(writeindex);
 
-        // restore
-        params->SetRobotActiveJoints(robot);
+        robot->SetActiveDOFValues(originalvalues);
+
+        // jitter again for initial collision
+        if( planningutils::JitterActiveDOF(robot) == 0 ) {
+            RAVELOG_WARN("jitter failed\n");
+            return false;
+        }
         robot->GetActiveDOFValues(params->vinitialconfig);
-        robot->SetActiveDOFValues(params->vgoalconfig);
-
-        // jitter again for goal
-        if( planningutils::JitterActiveDOF(robot) == 0 ) {
-            RAVELOG_WARN("failed\n");
-            return false;
-        }
 
         PlannerBasePtr rrtplanner = RaveCreatePlanner(GetEnv(),_strRRTPlannerName);
         if( !rrtplanner ) {